Move useCustomerEffortScoreExitPageTracker and utils to @woocommerce/customer-effort-score
