+++ packages/js/customer-effort-score/src/constants.js
@@ -1,0 +1,1 @@
+export const ALLOW_TRACKING_OPTION_NAME = 'woocommerce_allow_tracking';
