+++ packages/js/customer-effort-score/src/index.ts
@@ -3,3 +3,6 @@ export * from './components/customer-feedback-simple';
 export * from './components/customer-feedback-modal';
 export * from './components/product-mvp-feedback-modal';
 export * from './components/feedback-modal';
+export * from './hooks/use-customer-effort-score-exit-page-tracker';
+export * from './utils/customer-effort-score-exit-page';
+export * from './constants';
