+++ plugins/woocommerce-admin/client/customer-effort-score-tracks/constants.ts
@@ -2,4 +2,3 @@ export const SHOWN_FOR_ACTIONS_OPTION_NAME =
 	'woocommerce_ces_shown_for_actions';
 export const ADMIN_INSTALL_TIMESTAMP_OPTION_NAME =
 	'woocommerce_admin_install_timestamp';
-export const ALLOW_TRACKING_OPTION_NAME = 'woocommerce_allow_tracking';
